+++ test/jest/helpers/harness.js
@@ -1,4 +1,5 @@
-import { MemoryRouter } from 'react-router-dom';
+import { Router } from 'react-router-dom';
+import { createMemoryHistory } from 'history';
 
 import { StripesContext } from '@folio/stripes-core/src/StripesContext';
 
@@ -7,13 +8,15 @@ import buildStripes from '../__mock__/stripesCore.mock';
 
 const STRIPES = buildStripes();
 
-const Harness = ({ stripes, children }) => (
+const defaultHistory = createMemoryHistory();
+
+const Harness = ({ stripes, children, history = defaultHistory }) => (
   <StripesContext.Provider value={stripes || STRIPES}>
-    <MemoryRouter>
+    <Router history={history}>
       <IntlProvider>
         {children}
       </IntlProvider>
-    </MemoryRouter>
+    </Router>
   </StripesContext.Provider>
 );