+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Remove the Ordered through EBSCO option from filters. (UIEH-1118)
 * Package record: Show a message that proxy or token value may take some time to update. (UIEH-1116)
 * Usage & analysis accordion focus issues. (UIEH-1110)
+* Settings > eHoldings | Apply baseline keyboard shortcuts. (UIEH-1121)
 
 ## [6.0.2] (https://github.com/folio-org/ui-eholdings/tree/v6.0.2) (2021-04-16)
 * Fix unable to change Knowledge base name. (UIEH-1122)